Found some broccoli seeds from last year.
I like the sprouting kind of broccoli that makes side shoots after you
cut the main head.  I'm too lazy to rip it out of the garden, and
broccoli flowers are delicious.  Since I'm going to let it go to seed
anyway, I might as well get some extra sprouts on the side.

Signed-off-by: Sarah Sharp <sarah@thesharps.us>